Add functionality to basic properties.
GettableProperty
Property
Bool
String
StringChoice
Integer
Float
Complex
DateTime
Array
Vector3
Vector2
Color
